refactor(popup/Settings/Other): extract into components
